wip: documented and moved `theme::Loader::read_names` to `helix_loader::read_toml_names`
